ai-portfolio/ai-projects/projects/clinical-trial-eligibility-screener/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langfuse>=2.0.0
streamlit>=1.35.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
numpy>=1.24.0